authorization implemented

+++ src/api/playlistsongs/handler.js
@@ -1,4 +1,5 @@
 const ClientError = require('../../exceptions/ClientError')
+const NotFoundError = require('../../exceptions/NotFoundError')
 
 class PlaylistsongsHandler {
   constructor (playlistsongsService, playlistsService, validator) {
@@ -7,14 +8,20 @@ class PlaylistsongsHandler {
     this._validator = validator
 
     this.addPlaylistsongHandler = this.addPlaylistsongHandler.bind(this)
+    this.getPlaylistsongHandler = this.getPlaylistsongHandler.bind(this)
     this.deletePlaylistsongHandler = this.deletePlaylistsongHandler.bind(this)
   }
 
   async addPlaylistsongHandler (request, h) {
     try {
-      this._validator.validatePlaylistsongPayload(request.payload)
       const { id: credentialId } = request.auth.credentials
-      const { playlistId, songId } = request.payload
+      const { playlistId, any } = request.params
+      if (any !== 'songs') {
+        throw new NotFoundError('Resource not found')
+      }
+
+      this._validator.validatePlaylistsongPayload(request.payload)
+      const { songId } = request.payload
 
       await this._playlistsService.verifyPlaylistsongAccess(playlistId, credentialId)
       const playlistsongId = await this._playlistsongsService.addPlaylistsong(playlistId, songId)
@@ -49,11 +56,52 @@ class PlaylistsongsHandler {
     }
   }
 
+  async getPlaylistsongHandler (request, h) {
+    try {
+      const { id: credentialId } = request.auth.credentials
+      const { playlistId, any } = request.params
+      if (any !== 'songs') {
+        throw new NotFoundError('Resource not found')
+      }
+
+      await this._playlistsService.verifyPlaylistsongAccess(playlistId, credentialId)
+      const songs = await this._playlistsongsService.getPlaylistsong(playlistId)
+
+      return {
+        status: 'success',
+        data: {
+          songs
+        }
+      }
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        })
+        response.code(error.statusCode)
+        return response
+      }
+      const response = h.response({
+        status: 'error',
+        message: 'Sorry, there is a failure on our server.'
+      })
+      response.code(500)
+      console.error(error)
+      return response
+    }
+  }
+
   async deletePlaylistsongHandler (request, h) {
     try {
-      this._validator.validatePlaylistsongPayload(request.payload)
       const { id: credentialId } = request.auth.credentials
-      const { playlistId, songId } = request.payload
+      const { playlistId, any } = request.params
+      if (any !== 'songs') {
+        throw new NotFoundError('Resource not found')
+      }
+
+      this._validator.validatePlaylistsongPayload(request.payload)
+      const { songId } = request.payload
 
       await this._playlistsService.verifyPlaylistsongAccess(playlistId, credentialId)
       await this._playlistsongsService.deletePlaylistsong(playlistId, songId)
